rgb-image-segmentation/requirements.txt
numpy==1.21.0
opencv-python==4.5.3.20210927
torch==1.9.0
torchvision==0.10.0
matplotlib==3.4.3
scikit-learn==0.24.2
Pillow==8.4.0
PyYAML==5.4.1